+++ src/idlib/Heap.h
@@ -228,7 +228,7 @@ void		Mem_Free16( void *ptr );
 inline void *Mem_StackAlloc( const int size ) { return( _alloca( size ) ); }
 inline void *Mem_StackAlloc16( const int size ) { 
 	byte *addr = ( byte * )_alloca( size + 15 );
-	addr = ( byte * )( ( int )( addr + 15 ) & 0xfffffff0 );
+	addr = ( byte * )( ( intptr_t )( addr + 15 ) & 0xfffffff0 );
 	return( ( void * )addr ); 
 }
 
@@ -378,7 +378,7 @@ type *idBlockAlloc<type,blockSize,memoryTag>::Alloc( void ) {
 
 template<class type, int blockSize, byte memoryTag>
 void idBlockAlloc<type,blockSize,memoryTag>::Free( type *t ) {
-	element_t *element = (element_t *)( ( (unsigned char *) t ) - ( (int) &((element_t *)0)->t ) );
+	element_t *element = (element_t *)( ( (unsigned char *) t ) - ( (intptr_t) &((element_t *)0)->t ) );
 	element->next = free;
 	free = element;
 	active--;
